Add cc_emit.c/h files and move contents from cc_core.c
