+++ makefile
@@ -26,12 +26,13 @@ all: bin/M2-Planet
 .NOTPARALLEL:
 M2-Planet: bin/M2-Planet
 
-bin/M2-Planet: bin test/results cc.h cc_reader.c cc_strings.c cc_types.c cc_core.c cc.c cc_globals.c cc_globals.h cc_macro.c | bin
+bin/M2-Planet: bin test/results cc.h cc_reader.c cc_strings.c cc_types.c cc_core.c cc_emit.c cc_emit.h cc.c cc_globals.c cc_globals.h cc_macro.c | bin
 	$(CC) $(CFLAGS) \
 	M2libc/bootstrappable.c \
 	cc_reader.c \
 	cc_strings.c \
 	cc_types.c \
+	cc_emit.c \
 	cc_core.c \
 	cc_macro.c \
 	cc.c \
@@ -48,6 +49,7 @@ bin/M2-minimal: bin test/results cc.h cc_reader.c cc_strings.c cc_types.c cc_cor
 	cc_reader.c \
 	cc_strings.c \
 	cc_types.c \
+	cc_emit.c \
 	cc_core.c \
 	cc-minimal.c \
 	cc.h \
