+++ test/test.answers
@@ -406,10 +406,10 @@ d8c78b124a00fcd4f1b59d4b9ff88e5e87814b0012a730de60a01e0b91992901  test/results/t
 9cf352e597e18dd6f4c83453ad5627e86bd105582059d6c799a673e2b6676747  test/results/test0106-riscv32-binary
 b2af0a8d98e2ce0bf92e9c1a1fc7aedf30225fbd90a91c8ec7fb1185f6d0ad68  test/results/test0106-riscv64-binary
 dcfd0feda4f042fe8f589776c391a8ac676306d3e68222664bdadb023caf44ae  test/results/test0106-x86-binary
-1147c1d2a57ab7f0c3c79c73c03a2a804a5a16e8fa40f26ca77b693d9d757dca  test/results/test1000-aarch64-binary
-5517bb267c6825190b2d709f5b522aba1b10c6138b81645003971cf149450284  test/results/test1000-amd64-binary
-1dc859f04b234b2748e469d950dbca660e0499746c95fbdb8be0d0b110605cd7  test/results/test1000-armv7l-binary
-4b799e6ba474e760e8b9eafdad31c4e82321970c97cb4380d16ed731e2af2c4f  test/results/test1000-knight-posix-binary
-58c9efb8f758748255c28ddb9e6d211d96aa69a2c260b09b1eab56d4c1ffe7d8  test/results/test1000-riscv32-binary
-019fdf5eac8db1a9a609ae5af6e2933cef4d30a685d85f4536b97a007cae8b97  test/results/test1000-riscv64-binary
-5ea195d63e5e267edc96b8f8c5b60d08467ea5885e8bcd2af023219525b7b42d  test/results/test1000-x86-binary
+1e94ce7bb904abce7cda0cdf0af599b77967507eae1ce33d9f624612b3586b7b  test/results/test1000-aarch64-binary
+811913d704be89e532321f9fa4201b51e0acaf06f6708396004c270392122520  test/results/test1000-amd64-binary
+dd2e0be3076a875807529b0a7f05bf7cdd52202defdde7e7863bf9510bf69b05  test/results/test1000-armv7l-binary
+7401bd52f131bfec01943d7a579a555362b76baa0a18700758d54fa62c16180f  test/results/test1000-knight-posix-binary
+ab5a965a0ac029604012edc8bfd91b356ae67bb2e3df237054c68d3200f552f9  test/results/test1000-riscv32-binary
+680993e243f74b12b0c06b93825f6fc2154e47760f2a822a76806c68da5faaaa  test/results/test1000-riscv64-binary
+797e8b3b8d0f41556e263826a4b42e731bb0ddf712a430a40eec02644727d5b9  test/results/test1000-x86-binary
